pyproject.toml
[tool.poetry]
name = "geopephub"
version = "0.1.0"
description = "Pipeline for processing GEO data and uploading it to the PEPHub"
authors = ["Oleksandr Khoroshevskyi <khorosh@virginia.edu>"]
readme = "README.md"
homepage = "https://github.com/pepkit/geoephub"
documentation = "https://github.com/pepkit/geoephub"
repository = "https://github.com/pepkit/geoephub"
license = "BSD-2-Clause license"
exclude = []
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD-2-Clause license",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]

[tool.poetry.dependencies]
python = "^3.8"
geofetch = "^0.12.7"
pepdbagent= "^0.11.1"
#pepdbagent= { git = "https://github.com/pepkit/pepdbagent.git", branch = "dev" }
SQLAlchemy = ">=2.0.0,<2.1.0"
logmuse= "^0.2.7"
coloredlogs= "^15.0.1"
peppy= "^0.40.7"
pephubclient= "^0.4.4"
#pephubclient= { git = "https://github.com/pepkit/pephubclient.git", branch = "dev" }
pydantic= "^2.5.0"
hypothesis= "^6.112.0"
python-dotenv= "^1.0.1"
typer= "^0.9.0"
boto3="^1.34.29"
botocore="^1.34.29"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-version-plugin]
source = "init"

[tool.poetry.scripts]
geopephub = "geopephub.main:main"
